[REFACTOR] 검색 공고 패키징 수정
