+++ data/src/main/java/com/terning/data/mapper/search/SearchAnnouncementListMapper.kt
@@ -1,7 +1,7 @@
 package com.terning.data.mapper.search
 
 import com.terning.data.dto.response.SearchAnnouncementResponseDto
-import com.terning.domain.entity.response.SearchAnnouncement
+import com.terning.domain.entity.search.SearchAnnouncement
 
 fun SearchAnnouncementResponseDto.toSearchAnnouncementList(): List<SearchAnnouncement> {
     return announcements.map {
