+++ feature/src/main/java/com/terning/feature/search/search/SearchRoute.kt
@@ -28,7 +28,7 @@ import com.terning.core.designsystem.theme.Grey100
 import com.terning.core.designsystem.theme.TerningTheme
 import com.terning.core.extension.noRippleClickable
 import com.terning.core.state.UiState
-import com.terning.domain.entity.response.SearchAnnouncement
+import com.terning.domain.entity.search.SearchAnnouncement
 import com.terning.feature.R
 import com.terning.feature.search.search.component.ImageSlider
 import com.terning.feature.search.search.component.InternListType
