+++ feature/src/main/java/com/terning/feature/search/search/model/SearchViewsListState.kt
@@ -1,6 +1,6 @@
 package com.terning.feature.search.search.model
 
-import com.terning.domain.entity.response.SearchAnnouncement
+import com.terning.domain.entity.search.SearchAnnouncement
 import com.terning.core.state.UiState
 
 data class SearchViewsListState(
