+++ feature/src/main/java/com/terning/feature/search/searchprocess/models/SearchResultState.kt
@@ -1,8 +1,8 @@
 package com.terning.feature.search.searchprocess.models
 
 import com.terning.core.state.UiState
-import com.terning.domain.entity.response.SearchResultModel
+import com.terning.domain.entity.search.SearchResult
 
 data class SearchResultState(
-    var searchListState: UiState<List<SearchResultModel>> = UiState.Loading,
+    var searchListState: UiState<List<SearchResult>> = UiState.Loading,
 )
